// BDH.Rhino.Web.FrontEnd/src/app/modules/urban-solve/models/DataTransferObjects/TypologyOpbrensten.ts
export interface TypologyOpbrensten {
    typologyName: string;
    builtUnits: number;
    livingsPerUnit: number;
    totalLivings: number;

    grossIncomePerUnit: number;
    percentSocial: number;

    maxRentPriceSocial: number;
    rentFactorWozSocial: number;
    maxGrossIncomeSocial: number;
    builtSocial: number;
    profitSocial: number;
    builtFreeLivings: number;
    profitFreeLivings: number;
    totalProfit: number;
}